Made the plots simpler and added simple regression
I removed the plots we had made before and replaced them with all the simple plots that I think are relevant for the presentation.

I also changed the import of the data so we only use the data from 2000 to 2022, this way the modelling part becomes easier.

Lastly, I made some simple regression